# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Gappd Social API"
  description: "Used to manage core social features: users, posts, comments, events, photos"
  version: "0.0.1"
host: "gappd-backend.onrender.com" # During development my API is served on localhost (will change when we go to prod)
schemes:
  - "https"
consumes:
  - "application/json" # receives json data
produces:
  - "application/json" # produces json data
securityDefinitions: # Describing how to add token auth to the endpoints
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths: # Information about our routes
...
  /posts:
    post:
      tags:
        - Posts
      summary: "Create Post"
      description: "Create a post with optional photos; accepts uploaded files and/or existing photo_ids"
      security:
        - bearerAuth: []
      consumes:
        - "multipart/form-data"
      parameters:
        - in: "formData"
          name: "caption"
          type: "string"
          required: false
        - in: "formData"
          name: "location"
          type: "string"
          required: false
        - in: "formData"
          name: "photo_ids"
          type: "array"
          items:
            type: "string"
          collectionFormat: "multi"
          required: false
        - in: "formData"
          name: "files"
          type: "array"
          items:
            type: "file"
          collectionFormat: "multi"
          required: false
      responses:
        201:
          description: "Creation Successful"
          schema:
            $ref: "#/definitions/PostResponse"
  /posts/search:
    get:
      tags:
        - Posts
      summary: "Search posts by caption keywords"
      description: "Endpoint to receive an array of posts matching a caption keyword with pagination"
      parameters:
        - in: "query"
          name: "query_params"
          type: "string"
          required: true
        - in: "query"
          name: "page"
          type: "integer"
          required: false
      responses:
        200:
          description: "Successfully Retrieved Posts"
          schema:
            $ref: "#/definitions/PostsPageResponse"
  /posts/{post_id}:
    get:
      tags:
        - Posts
      summary: "Get a single post"
      parameters:
        - in: "path"
          name: "post_id"
          type: "integer"
          required: true
      responses:
        200:
          description: "Successfully Retrieved Post"
          schema:
            $ref: "#/definitions/PostResponse"

    put:
      tags:
        - Posts
      summary: "Update Post"
      description: "Update a post; requires token"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "post_id"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          description: "Post Updates"
          required: true
          schema:
            $ref: "#/definitions/PostUpdate"
      responses:
        200:
          description: "Update Successful"
          schema:
            $ref: "#/definitions/PostResponse"

    delete:
      tags:
        - Posts
      summary: "Delete a post"
      description: "Delete a post; requires token"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "post_id"
          type: "integer"
          required: true
      responses:
        200:
          description: "Deleted Post"
          schema:
            type: object
            properties:
              message:
                type: string
  /posts/by-user/{user_id}:
    get:
      tags:
        - Posts
      summary: "View all posts of a user"
      parameters:
        - in: "path"
          name: "user_id"
          type: "integer"
          required: true
        - in: "query"
          name: "page"
          type: "integer"
          required: false
      responses:
        200:
          description: "Successfully Retrieved User Posts"
          schema:
            $ref: "#/definitions/PostsPageResponse"
  /posts/feed:
    get:
      tags:
        - Posts
      summary: "View feed (from followed users)"
      security:
        - bearerAuth: []
      parameters:
        - in: "query"
          name: "page"
          type: "integer"
          required: false
      responses:
        200:
          description: "Successfully Retrieved Feed"
          schema:
            $ref: "#/definitions/PostsPageResponse"
  /posts/{post_id}/like:
    post:
      tags:
        - PostLikes
      summary: "Like a post"
      description: "Like a post, requires token"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "post_id"
          type: "integer"
          required: true
      responses:
        201:
          description: "Liked Post"
          schema:
            type: object
            properties:
              message:
                type: string

    delete:
      tags:
        - PostLikes
      summary: "Unlike a post"
      description: "Unlike a post, requires token"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "post_id"
          type: "integer"
          required: true
      responses:
        200:
          description: "Unliked Post"
          schema:
            type: object
            properties:
              message:
                type: string
  /posts/{post_id}/likes:
    get:
      tags:
        - PostLikes
      summary: "List users who liked a post"
      parameters:
        - in: "path"
          name: "post_id"
          type: "integer"
          required: true
        - in: "query"
          name: "page"
          type: "integer"
          required: false
      responses:
        200:
          description: "Successfully Retrieved Likes"
          schema:
            $ref: "#/definitions/UsersPageResponse"

  #==================== Comments ======================
...
definitions: # Information about the shape of input and output json
...
  UserResponse:
    type: object
    properties:
      id:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      email:
        type: string
      username:
        type: string
      profile_photo_id:
        type: integer
  UsersPageResponse:
    type: object
    properties:
      items:
        type: array
        items:
          $ref: "#/definitions/UserResponse"
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      pages:
        type: integer
...
  PostUpdate:
    type: object
    properties:
      caption:
        type: string
      location:
        type: string
  PostResponse:
    type: object
    properties:
      id:
        type: integer
        example: 1
      user_id:
        type: integer
        example: 1
      caption:
        type: string
      location:
        type: string
      created_at:
        type: string
        format: date-time
  PostsPageResponse:
    type: object
    properties:
      items:
        type: array
        items:
          $ref: "#/definitions/PostResponse"
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      pages:
        type: integer

  #==================== Comments ====================
...